refactor(electron/player): extract audio + orchestration bridges from usePlayerMachine
usePlayerMachine.ts was a 544-LOC god-hook conflating 11 concerns. Split
audio I/O (singleton HTMLAudioElement, TTS fetch lifecycle, prefetch) into
playerAudioBridge.ts and cross-system orchestration (activeParagraph clears,
pageRequest writes, republish trigger) into playerOrchestrationBridge.ts.

Hook drops to 318 LOC and the two side-effect surfaces become independently
testable and replaceable. This is the Phase 2 baseline for the player-actor
restructure (.parity/2026-05-28-player-actor-restructure/PLAN.md) — Phase 3
swaps these bridges for fromCallback actors.

No behaviour change. 152 machine tests + 5 writePath tests pass unchanged.
